Move ThrowAllErrorHandler to core-project
